backend/mysite/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.5
pandas==2.2.3 